Refactored id's (from int to long)

+++ BlackJack.Entity/Hand.cs
@@ -5,8 +5,8 @@ namespace BlackJack.Entities
 	[Table("Hand")]
 	public class Hand : BaseEntity
 	{
-		public int CardId { get; set; }
-		public int PlayerInGameId { get; set; }
+		public long CardId { get; set; }
+		public long PlayerInGameId { get; set; }
 
 		[Write(false)]
 		public virtual Card Card { get; set; }
